+++ HISTORY.md
@@ -1,3 +1,8 @@
+unreleased
+==========
+
+  * Fix `Accept-Language` dropping a `q` value with whitespace after the `;`
+
 1.0.0 / 2024-08-31
 ==================